refactor: split core and Postgres test suites by behavior
