+++ test/core.test.ts
@@ -1,1199 +1,3 @@
-import { describe, expect, it } from 'vitest'
-import packageJson from '../package.json'
-import {
-  AuditEventType,
-  DefaultAuthService,
-  EMAIL_OTP_PROVIDER_ID,
-  OtpChannel,
-  PASSWORD_PROVIDER_ID,
-  PHONE_OTP_PROVIDER_ID,
-  createDefaultAuthPolicy,
-  getUniAuthAttributionNotice,
-  isUniAuthError,
-  SessionStatus,
-  UNIAUTH_ATTRIBUTION,
-  UniAuthErrorCode,
-  VerificationPurpose,
-  VerificationStatus,
-  addSeconds,
-  asVerificationId,
-  createHmacSecretHasher,
-  type EmailSender,
-  type ProviderIdentityAssertion,
-} from '../src'
-import { InMemoryAuthStore, createInMemoryAuthKit, StaticAuthProvider } from '../src/testing'
-
-interface PackageMetadata {
-  readonly name: string
-  readonly license: string
-  readonly author: {
-    readonly name: string
-    readonly email: string
-  }
-}
-
-function formatPackageLicenseName(license: string): string {
-  return license
-    .replace(/^PolyForm-/, 'PolyForm ')
-    .replace(/-(\d+\.\d+\.\d+)$/, ' License $1')
-    .replaceAll('-', ' ')
-}
-
-const packageMetadata = packageJson as PackageMetadata
-const packageLicense = formatPackageLicenseName(packageMetadata.license)
-const now = new Date('2026-01-01T00:00:00.000Z')
-
-function assertion(input: Partial<ProviderIdentityAssertion> = {}): ProviderIdentityAssertion {
-  return {
-    provider: input.provider ?? 'email',
-    providerUserId: input.providerUserId ?? 'alice',
-    email: input.email ?? 'Alice@Example.com',
-    emailVerified: input.emailVerified ?? true,
-    displayName: input.displayName ?? 'Alice',
-    ...(input.phone ? { phone: input.phone } : {}),
-    ...(input.phoneVerified !== undefined ? { phoneVerified: input.phoneVerified } : {}),
-  }
-}
-
-describe('DefaultAuthService', () => {
-  it('exports stable attribution metadata and an About/Legal notice helper', () => {
-    expect(UNIAUTH_ATTRIBUTION.packageName).toBe(packageMetadata.name)
-    expect(UNIAUTH_ATTRIBUTION.contactEmail).toBe(packageMetadata.author.email)
-    expect(UNIAUTH_ATTRIBUTION.license).toBe(packageLicense)
-    expect(getUniAuthAttributionNotice()).toBe(
-      `This product uses ${packageMetadata.name}. ${UNIAUTH_ATTRIBUTION.copyright} License: ${packageLicense}. Licensing contact: ${packageMetadata.author.email}.`,
-    )
-    expect(
-      getUniAuthAttributionNotice({
-        includeContact: false,
-        includeLicense: false,
-        productName: 'Example App',
-      }),
-    ).toBe(`Example App uses ${packageMetadata.name}. ${UNIAUTH_ATTRIBUTION.copyright}`)
-  })
-
-  it('creates a local user, identity, and session for a new sign-in', async () => {
-    const { service, store } = createInMemoryAuthKit()
-
-    const result = await service.signIn({ assertion: assertion(), now })
-
-    expect(result.isNewUser).toBe(true)
-    expect(result.isNewIdentity).toBe(true)
-    expect(result.user.email).toBe('alice@example.com')
-    expect(result.identity.email).toBe('alice@example.com')
-    expect(result.session.status).toBe(SessionStatus.Active)
-    expect(result.sessionToken).toBeTypeOf('string')
-    expect(result.sessionToken).not.toBe(result.session.id)
-    expect(result.session.tokenHash).not.toBe(result.sessionToken)
-    expect(await service.getUser(result.user.id)).toBe(result.user)
-    expect(await service.resolveSession({ sessionToken: result.sessionToken, now })).toBe(
-      result.session,
-    )
-    expect(store.listUsers()).toHaveLength(1)
-    expect(store.listIdentities()).toHaveLength(1)
-    expect(store.listSessions()).toHaveLength(1)
-    expect(store.listAuditEvents().map((event) => event.type)).toContain(
-      AuditEventType.SessionCreated,
-    )
-  })
-
-  it('touches active sessions without rewinding last seen activity', async () => {
-    const { service } = createInMemoryAuthKit()
-    const result = await service.signIn({ assertion: assertion(), now })
-    const touchedAt = addSeconds(now, 60)
-    const touched = await service.touchSession({
-      sessionId: result.session.id,
-      now: touchedAt,
-    })
-
-    expect(touched.lastSeenAt).toEqual(touchedAt)
-    expect(
-      await service.touchSession({
-        sessionId: result.session.id,
-        now: addSeconds(now, 30),
-      }),
-    ).toBe(touched)
-  })
-
-  it('reads local credentials and verifications through the public service surface', async () => {
-    const { service } = createInMemoryAuthKit()
-    const signedIn = await service.signIn({
-      assertion: assertion({
-        provider: 'email',
-        providerUserId: 'credential-reader',
-        email: 'credential-reader@example.com',
-        emailVerified: true,
-      }),
-      now,
-    })
-    const credential = await service.setPassword({
-      userId: signedIn.user.id,
-      email: 'credential-reader@example.com',
-      password: 'secret-password',
-      now,
-    })
-    const createdVerification = await service.createVerification({
-      purpose: VerificationPurpose.SignIn,
-      target: 'credential-reader@example.com',
-      secret: '123456',
-      now,
-    })
-
-    expect(await service.getUserCredentials(signedIn.user.id)).toEqual([credential])
-    expect(await service.getVerification(createdVerification.verification.id)).toEqual(
-      createdVerification.verification,
-    )
-    expect(await service.getAccountSecuritySnapshot(signedIn.user.id)).toEqual({
-      user: {
-        id: signedIn.user.id,
-        email: 'credential-reader@example.com',
-        displayName: 'Alice',
-        createdAt: signedIn.user.createdAt,
-        updatedAt: signedIn.user.updatedAt,
-      },
-      identities: [
-        {
-          id: signedIn.identity.id,
-          provider: signedIn.identity.provider,
-          status: signedIn.identity.status,
-          email: 'credential-reader@example.com',
-          emailVerified: true,
-          createdAt: signedIn.identity.createdAt,
-          updatedAt: signedIn.identity.updatedAt,
-        },
-        {
-          provider: PASSWORD_PROVIDER_ID,
-          status: 'active',
-          email: 'credential-reader@example.com',
-          emailVerified: true,
-          id: expect.any(String),
-          createdAt: credential.createdAt,
-          updatedAt: credential.updatedAt,
-        },
-      ],
-      credentials: [
-        {
-          id: credential.id,
-          type: credential.type,
-          subject: credential.subject,
-          createdAt: credential.createdAt,
-          updatedAt: credential.updatedAt,
-        },
-      ],
-      sessions: [
-        {
-          id: signedIn.session.id,
-          status: signedIn.session.status,
-          createdAt: signedIn.session.createdAt,
-          expiresAt: signedIn.session.expiresAt,
-        },
-      ],
-    })
-  })
-
-  it('bulk-revokes active user sessions while optionally keeping one session active', async () => {
-    const { service, store } = createInMemoryAuthKit()
-    const signedIn = await service.signIn({ assertion: assertion(), now })
-    const second = await service.createSession({
-      userId: signedIn.user.id,
-      now: addSeconds(now, 10),
-    })
-    const third = await service.createSession({
-      userId: signedIn.user.id,
-      now: addSeconds(now, 20),
-    })
-
-    const result = await service.revokeUserSessions({
-      userId: signedIn.user.id,
-      exceptSessionId: signedIn.session.id,
-      now: addSeconds(now, 30),
-    })
-
-    expect(result).toEqual({
-      userId: signedIn.user.id,
-      revokedSessionIds: [second.session.id, third.session.id],
-    })
-    expect(await service.getUserSessions(signedIn.user.id)).toMatchObject([
-      { id: signedIn.session.id, status: SessionStatus.Active },
-      { id: second.session.id, status: SessionStatus.Revoked },
-      { id: third.session.id, status: SessionStatus.Revoked },
-    ])
-    expect(
-      store
-        .listAuditEvents()
-        .filter((event) => event.type === AuditEventType.SessionRevoked)
-        .map((event) => event.sessionId),
-    ).toEqual([second.session.id, third.session.id])
-  })
-
-  it('honors explicit zero TTL options in the in-memory testing kit', async () => {
-    const { service } = createInMemoryAuthKit({
-      clock: { now: () => now },
-      sessionTtlSeconds: 0,
-      verificationTtlSeconds: 0,
-    })
-
-    const result = await service.signIn({ assertion: assertion() })
-    const verification = await service.createVerification({
-      purpose: VerificationPurpose.SignIn,
-      target: 'ttl@example.com',
-    })
-
-    expect(result.session.expiresAt).toEqual(now)
-    expect(verification.verification.expiresAt).toEqual(now)
-  })
-
-  it('rejects invalid session tokens, TTLs, and expiration inputs', async () => {
-    const { service } = createInMemoryAuthKit()
-    const signedIn = await service.signIn({ assertion: assertion(), now })
-    const expired = await service.createSession({
-      userId: signedIn.user.id,
-      expiresAt: now,
-      now,
-    })
-
-    await expect(service.resolveSession({ sessionToken: '   ', now })).rejects.toMatchObject({
-      code: UniAuthErrorCode.InvalidInput,
-    })
-    await expect(service.resolveSession({ sessionToken: 'missing', now })).rejects.toMatchObject({
-      code: UniAuthErrorCode.SessionNotFound,
-    })
-    await expect(
-      service.resolveSession({ sessionToken: expired.sessionToken, now }),
-    ).rejects.toMatchObject({
-      code: UniAuthErrorCode.SessionNotFound,
-    })
-    await expect(
-      service.touchSession({
-        sessionId: expired.session.id,
-        now,
-      }),
-    ).rejects.toMatchObject({
-      code: UniAuthErrorCode.SessionNotFound,
-    })
-    await expect(
-      service.createSession({
-        userId: signedIn.user.id,
-        expiresAt: addSeconds(now, -1),
-        now,
-      }),
-    ).rejects.toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    await expect(
-      service.createSession({
-        userId: signedIn.user.id,
-        expiresAt: new Date('invalid'),
-        now,
-      }),
-    ).rejects.toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    await expect(
-      service.createSession({
-        userId: signedIn.user.id,
-        now: new Date('invalid'),
-      }),
-    ).rejects.toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-
-    await expect(
-      createInMemoryAuthKit({ sessionTtlSeconds: -1 }).service.signIn({
-        assertion: assertion(),
-        now,
-      }),
-    ).rejects.toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    await expect(
-      createInMemoryAuthKit({ sessionTtlSeconds: Number.NaN }).service.signIn({
-        assertion: assertion(),
-        now,
-      }),
-    ).rejects.toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    await expect(
-      service.touchSession({
-        sessionId: signedIn.session.id,
-        now: new Date('invalid'),
-      }),
-    ).rejects.toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-  })
-
-  it('rejects invalid verification TTL and creation time inputs', async () => {
-    const { service } = createInMemoryAuthKit()
-
-    await expect(
-      service.createVerification({
-        purpose: VerificationPurpose.SignIn,
-        target: 'ttl@example.com',
-        ttlSeconds: -1,
-        now,
-      }),
-    ).rejects.toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    await expect(
-      service.createVerification({
-        purpose: VerificationPurpose.SignIn,
-        target: 'nan@example.com',
-        ttlSeconds: Number.NaN,
-        now,
-      }),
-    ).rejects.toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    await expect(
-      service.createVerification({
-        purpose: VerificationPurpose.SignIn,
-        target: 'invalid-now@example.com',
-        now: new Date('invalid'),
-      }),
-    ).rejects.toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-
-    await expect(
-      createInMemoryAuthKit({ verificationTtlSeconds: -1 }).service.createVerification({
-        purpose: VerificationPurpose.SignIn,
-        target: 'runtime-negative@example.com',
-        now,
-      }),
-    ).rejects.toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    await expect(
-      createInMemoryAuthKit({ verificationTtlSeconds: Number.NaN }).service.createVerification({
-        purpose: VerificationPurpose.SignIn,
-        target: 'runtime-nan@example.com',
-        now,
-      }),
-    ).rejects.toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-  })
-
-  it('uses the runtime clock when resolving sessions and finishing OTP sign-in', async () => {
-    const { service } = createInMemoryAuthKit({
-      clock: { now: () => now },
-    })
-    const signedIn = await service.signIn({
-      assertion: assertion({
-        providerUserId: 'clock-session',
-        email: 'clock-session@example.com',
-      }),
-    })
-
-    expect(await service.resolveSession({ sessionToken: signedIn.sessionToken })).toBe(
-      signedIn.session,
-    )
-    expect(await service.touchSession({ sessionId: signedIn.session.id })).toMatchObject({
-      id: signedIn.session.id,
-      lastSeenAt: now,
-    })
-
-    const challenge = await service.startOtpChallenge({
-      purpose: VerificationPurpose.SignIn,
-      channel: OtpChannel.Email,
-      target: 'clock-otp@example.com',
-      secret: '123456',
-    })
-    const finished = await service.finishOtpSignIn({
-      verificationId: challenge.verificationId,
-      secret: '123456',
-      channel: OtpChannel.Email,
-    })
-
-    expect(finished.session.status).toBe(SessionStatus.Active)
-  })
-
-  it('uses the runtime clock when bulk-revoking user sessions', async () => {
-    const { service } = createInMemoryAuthKit({
-      clock: { now: () => now },
-    })
-    const signedIn = await service.signIn({ assertion: assertion() })
-    const second = await service.createSession({
-      userId: signedIn.user.id,
-    })
-
-    expect(await service.revokeUserSessions({ userId: signedIn.user.id })).toEqual({
-      userId: signedIn.user.id,
-      revokedSessionIds: [signedIn.session.id, second.session.id],
-    })
-    expect(await service.getUserSessions(signedIn.user.id)).toMatchObject([
-      { id: signedIn.session.id, status: SessionStatus.Revoked, revokedAt: now },
-      { id: second.session.id, status: SessionStatus.Revoked, revokedAt: now },
-    ])
-  })
-
-  it('uses exact provider identity match before any profile matching', async () => {
-    const { service, store } = createInMemoryAuthKit()
-
-    const first = await service.signIn({ assertion: assertion(), now })
-    const second = await service.signIn({
-      assertion: assertion({ email: 'different@example.com' }),
-      now,
-    })
-
-    expect(second.isNewUser).toBe(false)
-    expect(second.isNewIdentity).toBe(false)
-    expect(second.user.id).toBe(first.user.id)
-    expect(store.listUsers()).toHaveLength(1)
-    expect(store.listIdentities()).toHaveLength(1)
-    expect(store.listSessions()).toHaveLength(2)
-  })
-
-  it('does not silently merge users by verified email under the default policy', async () => {
-    const { service, store } = createInMemoryAuthKit()
-
-    const emailUser = await service.signIn({ assertion: assertion(), now })
-    const oauthUser = await service.signIn({
-      assertion: assertion({ provider: 'oauth', providerUserId: 'oauth-alice' }),
-      now,
-    })
-
-    expect(oauthUser.isNewUser).toBe(true)
-    expect(oauthUser.user.id).not.toBe(emailUser.user.id)
-    expect(store.listUsers()).toHaveLength(2)
-    expect(store.listIdentities()).toHaveLength(2)
-  })
-
-  it('auto-links only when the policy explicitly allows it', async () => {
-    const { service, store } = createInMemoryAuthKit({
-      policy: createDefaultAuthPolicy({ allowAutoLink: true }),
-    })
-
-    const emailUser = await service.signIn({ assertion: assertion(), now })
-    const oauthUser = await service.signIn({
-      assertion: assertion({ provider: 'oauth', providerUserId: 'oauth-alice' }),
-      now,
-    })
-
-    expect(oauthUser.isNewUser).toBe(false)
-    expect(oauthUser.isNewIdentity).toBe(true)
-    expect(oauthUser.user.id).toBe(emailUser.user.id)
-    expect(store.listUsers()).toHaveLength(1)
-    expect(store.listIdentities()).toHaveLength(2)
-  })
-
-  it('rejects unlinking the last active identity', async () => {
-    const { service } = createInMemoryAuthKit()
-    const result = await service.signIn({ assertion: assertion(), now })
-
-    const error = await service
-      .unlink({ userId: result.user.id, identityId: result.identity.id, now })
-      .catch((caught: unknown) => caught)
-
-    expect(error).toMatchObject({ code: UniAuthErrorCode.LastIdentity })
-  })
-
-  it('rejects linking an identity already attached to another user without leaking ownership', async () => {
-    const { service } = createInMemoryAuthKit()
-    const first = await service.signIn({ assertion: assertion(), now })
-    const second = await service.signIn({
-      assertion: assertion({
-        provider: 'email',
-        providerUserId: 'bob',
-        email: 'bob@example.com',
-        displayName: 'Bob',
-      }),
-      now,
-    })
-
-    const error = await service
-      .link({
-        userId: second.user.id,
-        assertion: assertion({
-          provider: first.identity.provider,
-          providerUserId: first.identity.providerUserId,
-        }),
-        now,
-      })
-      .catch((caught: unknown) => caught)
-
-    expect(isUniAuthError(error)).toBe(true)
-
-    if (!isUniAuthError(error)) {
-      throw new Error('Expected a UniAuthError.')
-    }
-
-    expect(error.code).toBe(UniAuthErrorCode.IdentityAlreadyLinked)
-    expect(error.message).not.toContain('another user')
-    expect(error.message).not.toContain('account')
-  })
-
-  it('stores verification secrets only as hashes and consumes valid secrets once', async () => {
-    const { service, store } = createInMemoryAuthKit()
-
-    const created = await service.createVerification({
-      purpose: VerificationPurpose.SignIn,
-      target: 'Alice@Example.com',
-      secret: '123456',
-      now,
-    })
-
-    expect(created.secret).toBe('123456')
-    expect(created.verification.target).toBe('alice@example.com')
-    expect(created.verification.secretHash).not.toBe('123456')
-    expect(created.verification.secretHash).toMatch(/^scrypt:/)
-    expect(store.listVerifications()[0]?.secretHash).toBe(created.verification.secretHash)
-
-    const invalidSecretError = await service
-      .consumeVerification({
-        verificationId: created.verification.id,
-        secret: 'wrong',
-        now,
-      })
-      .catch((caught: unknown) => caught)
-
-    expect(invalidSecretError).toMatchObject({
-      code: UniAuthErrorCode.VerificationInvalidSecret,
-    })
-
-    const consumed = await service.consumeVerification({
-      verificationId: created.verification.id,
-      secret: '123456',
-      now,
-    })
-
-    expect(consumed.status).toBe(VerificationStatus.Consumed)
-    const consumedAgainError = await service
-      .consumeVerification({
-        verificationId: created.verification.id,
-        secret: '123456',
-        now,
-      })
-      .catch((caught: unknown) => caught)
-
-    expect(consumedAgainError).toMatchObject({
-      code: UniAuthErrorCode.VerificationConsumed,
-    })
-
-    const blankTargetError = await service
-      .createVerification({
-        purpose: VerificationPurpose.SignIn,
-        target: '   ',
-        secret: '123456',
-        now,
-      })
-      .catch((caught: unknown) => caught)
-
-    expect(blankTargetError).toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    expect(store.listVerifications()).toHaveLength(1)
-  })
-
-  it('uses a configured secret hasher for verification storage', async () => {
-    const { service, store } = createInMemoryAuthKit({
-      secretHasher: createHmacSecretHasher({ pepper: 'test-pepper' }),
-    })
-
-    const created = await service.createVerification({
-      purpose: VerificationPurpose.SignIn,
-      target: 'Alice@Example.com',
-      secret: '123456',
-      now,
-    })
-
-    expect(created.verification.secretHash).toMatch(/^hmac-sha256:/)
-    expect(store.listVerifications()[0]?.secretHash).toBe(created.verification.secretHash)
-
-    const consumed = await service.consumeVerification({
-      verificationId: created.verification.id,
-      secret: '123456',
-      now,
-    })
-
-    expect(consumed.status).toBe(VerificationStatus.Consumed)
-  })
-
-  it('starts and finishes email OTP sign-in without exposing account state', async () => {
-    const { emailSender, service, store } = createInMemoryAuthKit()
-
-    const started = await service.startOtpChallenge({
-      purpose: VerificationPurpose.SignIn,
-      channel: OtpChannel.Email,
-      target: ' Alice@Example.com ',
-      secret: '123456',
-      metadata: { requestId: 'req-1' },
-      now,
-    })
-
-    expect(started).toEqual({
-      verificationId: started.verificationId,
-      expiresAt: new Date('2026-01-01T00:10:00.000Z'),
-      delivery: 'email',
-    })
-    expect(started).not.toHaveProperty('isNewUser')
-    expect(started).not.toHaveProperty('user')
-
-    const [message] = emailSender.listMessages()
-    const [storedVerification] = store.listVerifications()
-
-    expect(message).toMatchObject({
-      to: 'alice@example.com',
-      subject: 'Your sign-in code',
-      text: 'Your sign-in code is 123456.',
-    })
-    expect(message?.metadata).toMatchObject({
-      verificationId: started.verificationId,
-      purpose: VerificationPurpose.SignIn,
-      delivery: 'email',
-    })
-    expect(storedVerification).toMatchObject({
-      id: started.verificationId,
-      purpose: VerificationPurpose.SignIn,
-      target: 'alice@example.com',
-      provider: EMAIL_OTP_PROVIDER_ID,
-      channel: OtpChannel.Email,
-      status: VerificationStatus.Pending,
-      metadata: {
-        requestId: 'req-1',
-      },
-    })
-    expect(storedVerification?.secretHash).not.toBe('123456')
-    expect(storedVerification?.secretHash).toMatch(/^scrypt:/)
-
-    const wrongSecret = await service
-      .finishOtpSignIn({
-        verificationId: started.verificationId,
-        secret: '000000',
-        channel: OtpChannel.Email,
-        now,
-      })
-      .catch((caught: unknown) => caught)
-
-    expect(wrongSecret).toMatchObject({
-      code: UniAuthErrorCode.VerificationInvalidSecret,
-    })
-    expect(store.listVerifications()[0]?.status).toBe(VerificationStatus.Pending)
-
-    const finished = await service.finishOtpSignIn({
-      verificationId: started.verificationId,
-      secret: '123456',
-      channel: OtpChannel.Email,
-      metadata: { flow: 'otp' },
-      now,
-    })
-
-    expect(finished.isNewUser).toBe(true)
-    expect(finished.identity.provider).toBe(EMAIL_OTP_PROVIDER_ID)
-    expect(finished.identity.providerUserId).toBe('alice@example.com')
-    expect(finished.identity.email).toBe('alice@example.com')
-    expect(finished.session.status).toBe(SessionStatus.Active)
-    expect(store.listVerifications()[0]?.status).toBe(VerificationStatus.Consumed)
-
-    const consumedAgain = await service
-      .finishOtpSignIn({
-        verificationId: started.verificationId,
-        secret: '123456',
-        channel: OtpChannel.Email,
-        now,
-      })
-      .catch((caught: unknown) => caught)
-
-    expect(consumedAgain).toMatchObject({
-      code: UniAuthErrorCode.VerificationConsumed,
-    })
-
-    const generated = await service.startOtpChallenge({
-      purpose: VerificationPurpose.SignIn,
-      channel: OtpChannel.Email,
-      target: 'bob@example.com',
-      ttlSeconds: 30,
-      now,
-    })
-    const generatedMessage = emailSender.listMessages()[1]
-    const generatedSecret = generatedMessage?.text.match(/\d{6}/)?.[0]
-
-    if (!generatedSecret) {
-      throw new Error('Expected generated OTP secret in the email message.')
-    }
-
-    const generatedFinished = await service.finishOtpSignIn({
-      verificationId: generated.verificationId,
-      secret: generatedSecret,
-      channel: OtpChannel.Email,
-      sessionExpiresAt: addSeconds(now, 60),
-      now,
-    })
-
-    expect(generated.delivery).toBe('email')
-    expect(generatedFinished.session.expiresAt).toEqual(addSeconds(now, 60))
-  })
-
-  it('keeps a pending OTP verification when app-owned delivery fails', async () => {
-    const store = new InMemoryAuthStore()
-    const failingEmailSender: EmailSender = {
-      sendEmail: async () => {
-        throw new Error('Delivery failed.')
-      },
-    }
-    const service = new DefaultAuthService({
-      repos: store,
-      emailSender: failingEmailSender,
-    })
-
-    const error = await service
-      .startOtpChallenge({
-        purpose: VerificationPurpose.SignIn,
-        channel: OtpChannel.Email,
-        target: 'delivery@example.com',
-        secret: '123456',
-        now,
-      })
-      .catch((caught: unknown) => caught)
-
-    expect(error).toBeInstanceOf(Error)
-    expect(store.listVerifications()).toHaveLength(1)
-    expect(store.listVerifications()[0]).toMatchObject({
-      purpose: VerificationPurpose.SignIn,
-      target: 'delivery@example.com',
-      provider: EMAIL_OTP_PROVIDER_ID,
-      channel: OtpChannel.Email,
-      status: VerificationStatus.Pending,
-    })
-    expect(store.listVerifications()[0]?.secretHash).not.toBe('123456')
-  })
-
-  it('reuses generic OTP challenges for email and phone sign-in channels', async () => {
-    const { emailSender, service, smsSender, store } = createInMemoryAuthKit()
-
-    const emailLinkChallenge = await service.startOtpChallenge({
-      purpose: VerificationPurpose.Link,
-      channel: OtpChannel.Email,
-      target: ' Link@Example.com ',
-      secret: '111111',
-      now,
-    })
-    const consumedEmailLink = await service.finishOtpChallenge({
-      verificationId: emailLinkChallenge.verificationId,
-      secret: '111111',
-      purpose: VerificationPurpose.Link,
-      channel: OtpChannel.Email,
-      now,
-    })
-
-    expect(emailLinkChallenge.delivery).toBe(OtpChannel.Email)
-    expect(consumedEmailLink.status).toBe(VerificationStatus.Consumed)
-    expect(emailSender.listMessages()[0]).toMatchObject({
-      to: 'link@example.com',
-      text: 'Your sign-in code is 111111.',
-    })
-
-    const clockFallbackChallenge = await service.startOtpChallenge({
-      purpose: VerificationPurpose.Link,
-      channel: OtpChannel.Email,
-      target: 'clock@example.com',
-      secret: '222222',
-    })
-    const clockFallbackConsumed = await service.finishOtpChallenge({
-      verificationId: clockFallbackChallenge.verificationId,
-      secret: '222222',
-      purpose: VerificationPurpose.Link,
-      channel: OtpChannel.Email,
-    })
-
-    expect(clockFallbackConsumed.status).toBe(VerificationStatus.Consumed)
-
-    const phoneChallenge = await service.startOtpChallenge({
-      purpose: VerificationPurpose.SignIn,
-      channel: OtpChannel.Phone,
-      target: ' +1 (555) 123-4567 ',
-      secret: '654321',
-      metadata: { requestId: 'req-phone' },
-      now,
-    })
-
-    expect(phoneChallenge).toEqual({
-      verificationId: phoneChallenge.verificationId,
-      expiresAt: new Date('2026-01-01T00:10:00.000Z'),
-      delivery: OtpChannel.Phone,
-    })
-    expect(smsSender.listMessages()[0]).toMatchObject({
-      to: '+15551234567',
-      text: 'Your sign-in code is 654321.',
-      metadata: {
-        verificationId: phoneChallenge.verificationId,
-        purpose: VerificationPurpose.SignIn,
-        delivery: OtpChannel.Phone,
-      },
-    })
-    expect(store.listVerifications()[2]).toMatchObject({
-      id: phoneChallenge.verificationId,
-      target: '+15551234567',
-      provider: PHONE_OTP_PROVIDER_ID,
-      channel: OtpChannel.Phone,
-      metadata: {
-        requestId: 'req-phone',
-      },
-    })
-
-    const phoneResult = await service.finishOtpSignIn({
-      verificationId: phoneChallenge.verificationId,
-      secret: '654321',
-      metadata: { flow: 'phone-otp' },
-      now,
-    })
-
-    expect(phoneResult.identity.provider).toBe(PHONE_OTP_PROVIDER_ID)
-    expect(phoneResult.identity.providerUserId).toBe('+15551234567')
-    expect(phoneResult.identity.phone).toBe('+15551234567')
-    expect(phoneResult.identity.phoneVerified).toBe(true)
-    expect(phoneResult.session.status).toBe(SessionStatus.Active)
-  })
-
-  it('rejects invalid email OTP sign-in starts and wrong verification purposes', async () => {
-    const serviceWithoutEmailSender = new DefaultAuthService({
-      repos: new InMemoryAuthStore(),
-    })
-
-    expect(
-      await serviceWithoutEmailSender
-        .startOtpChallenge({
-          purpose: VerificationPurpose.SignIn,
-          channel: OtpChannel.Email,
-          target: 'alice@example.com',
-          now,
-        })
-        .catch((caught: unknown) => caught),
-    ).toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    expect(
-      await serviceWithoutEmailSender
-        .startOtpChallenge({
-          purpose: VerificationPurpose.SignIn,
-          channel: OtpChannel.Email,
-          target: '   ',
-          now,
-        })
-        .catch((caught: unknown) => caught),
-    ).toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    expect(
-      await serviceWithoutEmailSender
-        .finishOtpSignIn({
-          verificationId: asVerificationId('missing'),
-          secret: '123456',
-          channel: OtpChannel.Email,
-          now,
-        })
-        .catch((caught: unknown) => caught),
-    ).toMatchObject({ code: UniAuthErrorCode.VerificationNotFound })
-
-    const { service, store } = createInMemoryAuthKit()
-    const linkVerification = await service.createVerification({
-      purpose: VerificationPurpose.Link,
-      target: 'alice@example.com',
-      secret: '123456',
-      now,
-    })
-    const wrongPurpose = await service
-      .finishOtpSignIn({
-        verificationId: linkVerification.verification.id,
-        secret: '123456',
-        channel: OtpChannel.Email,
-        now,
-      })
-      .catch((caught: unknown) => caught)
-
-    expect(wrongPurpose).toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    expect(store.listVerifications()[0]?.status).toBe(VerificationStatus.Pending)
-  })
-
-  it('rejects invalid generic OTP challenge usage without consuming secrets', async () => {
-    const serviceWithoutSmsSender = new DefaultAuthService({
-      repos: new InMemoryAuthStore(),
-    })
-
-    expect(
-      await serviceWithoutSmsSender
-        .startOtpChallenge({
-          purpose: VerificationPurpose.SignIn,
-          channel: OtpChannel.Phone,
-          target: '+15551234567',
-          now,
-        })
-        .catch((caught: unknown) => caught),
-    ).toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    expect(
-      await serviceWithoutSmsSender
-        .startOtpChallenge({
-          purpose: VerificationPurpose.SignIn,
-          // @ts-expect-error Unsupported channels are a runtime guard for untyped callers.
-          channel: 'push',
-          target: 'alice',
-          now,
-        })
-        .catch((caught: unknown) => caught),
-    ).toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    expect(
-      await serviceWithoutSmsSender
-        .startOtpChallenge({
-          purpose: VerificationPurpose.SignIn,
-          // @ts-expect-error Unsupported channels are a runtime guard for untyped callers.
-          channel: 'push',
-          target: '   ',
-          now,
-        })
-        .catch((caught: unknown) => caught),
-    ).toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    expect(
-      await serviceWithoutSmsSender
-        .startOtpChallenge({
-          purpose: VerificationPurpose.SignIn,
-          channel: OtpChannel.Phone,
-          target: '   ',
-          now,
-        })
-        .catch((caught: unknown) => caught),
-    ).toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-
-    const { service, store } = createInMemoryAuthKit()
-    const challenge = await service.startOtpChallenge({
-      purpose: VerificationPurpose.SignIn,
-      channel: OtpChannel.Email,
-      target: 'alice@example.com',
-      secret: '123456',
-      now,
-    })
-    const wrongChannel = await service
-      .finishOtpChallenge({
-        verificationId: challenge.verificationId,
-        secret: '123456',
-        purpose: VerificationPurpose.SignIn,
-        channel: OtpChannel.Phone,
-        now,
-      })
-      .catch((caught: unknown) => caught)
-
-    expect(wrongChannel).toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    expect(store.listVerifications()[0]?.status).toBe(VerificationStatus.Pending)
-
-    const rawVerification = await service.createVerification({
-      purpose: VerificationPurpose.SignIn,
-      target: 'raw@example.com',
-      secret: '123456',
-      now,
-    })
-    const notOtpChallenge = await service
-      .finishOtpChallenge({
-        verificationId: rawVerification.verification.id,
-        secret: '123456',
-        now,
-      })
-      .catch((caught: unknown) => caught)
-
-    expect(notOtpChallenge).toMatchObject({ code: UniAuthErrorCode.InvalidInput })
-    expect(store.listVerifications()[1]?.status).toBe(VerificationStatus.Pending)
-  })
-
-  it('requires explicit merge policy and moves identities only through mergeAccounts', async () => {
-    const deniedKit = createInMemoryAuthKit()
-    const deniedSource = await deniedKit.service.signIn({
-      assertion: assertion({ providerUserId: 'source', email: 'source@example.com' }),
-      now,
-    })
-    const deniedTarget = await deniedKit.service.signIn({
-      assertion: assertion({ providerUserId: 'target', email: 'target@example.com' }),
-      now,
-    })
-
-    const deniedMergeError = await deniedKit.service
-      .mergeAccounts({
-        sourceUserId: deniedSource.user.id,
-        targetUserId: deniedTarget.user.id,
-        reAuthenticatedAt: now,
-        now,
-      })
-      .catch((caught: unknown) => caught)
-
-    expect(deniedMergeError).toMatchObject({ code: UniAuthErrorCode.PolicyDenied })
-
-    const allowedKit = createInMemoryAuthKit({
-      policy: createDefaultAuthPolicy({ allowMergeAccounts: true, requireReAuthFor: [] }),
-    })
-    const source = await allowedKit.service.signIn({
-      assertion: assertion({ providerUserId: 'source', email: 'source@example.com' }),
-      now,
-    })
-    const target = await allowedKit.service.signIn({
-      assertion: assertion({ providerUserId: 'target', email: 'target@example.com' }),
-      now,
-    })
-
-    const merged = await allowedKit.service.mergeAccounts({
-      sourceUserId: source.user.id,
-      targetUserId: target.user.id,
-      now,
-    })
-
-    expect(merged.movedIdentityIds).toEqual([source.identity.id])
-    expect(merged.movedCredentialIds).toEqual([])
-    expect(merged.revokedSessionIds).toEqual([source.session.id])
-    expect(merged.sourceUser.disabledAt).toEqual(now)
-    expect(
-      allowedKit.store
-        .listIdentities()
-        .filter((identity) => identity.id === source.identity.id)
-        .map((identity) => identity.userId),
-    ).toEqual([target.user.id])
-    expect(
-      allowedKit.store
-        .listSessions()
-        .filter((session) => session.userId === source.user.id)
-        .map((session) => session.status),
-    ).toEqual([SessionStatus.Revoked])
-  })
-
-  it('moves password credentials on merge, supports idempotent retry, and keeps audit metadata secret-free', async () => {
-    const kit = createInMemoryAuthKit({
-      policy: createDefaultAuthPolicy({ allowMergeAccounts: true, requireReAuthFor: [] }),
-    })
-    const source = await kit.service.signIn({
-      assertion: assertion({ providerUserId: 'source-password', email: 'source@example.com' }),
-      now,
-    })
-    const target = await kit.service.signIn({
-      assertion: assertion({ providerUserId: 'target-password', email: 'target@example.com' }),
-      now,
-    })
-    const sourceCredential = await kit.service.setPassword({
-      userId: source.user.id,
-      email: 'source@example.com',
-      password: 'source-secret',
-      now,
-    })
-
-    const merged = await kit.service.mergeAccounts({
-      sourceUserId: source.user.id,
-      targetUserId: target.user.id,
-      reAuthenticatedAt: now,
-      now,
-    })
-
-    expect(merged.movedCredentialIds).toEqual([sourceCredential.id])
-    expect(merged.revokedSessionIds).toEqual([source.session.id])
-    expect(merged.movedIdentityIds).toEqual(expect.arrayContaining([source.identity.id]))
-    expect(
-      kit.store
-        .listCredentials()
-        .filter((credential) => credential.id === sourceCredential.id)
-        .map((credential) => credential.userId),
-    ).toEqual([target.user.id])
-
-    const passwordSignIn = await kit.service.signInWithPassword({
-      email: 'source@example.com',
-      password: 'source-secret',
-      now,
-    })
-
-    expect(passwordSignIn.user.id).toBe(target.user.id)
-
-    const retriedMerge = await kit.service.mergeAccounts({
-      sourceUserId: source.user.id,
-      targetUserId: target.user.id,
-      reAuthenticatedAt: now,
-      now,
-    })
-
-    expect(retriedMerge.movedIdentityIds).toEqual([])
-    expect(retriedMerge.movedCredentialIds).toEqual([])
-    expect(retriedMerge.revokedSessionIds).toEqual([])
-
-    const mergeEvents = kit.store
-      .listAuditEvents()
-      .filter((event) => event.type === AuditEventType.AccountsMerged)
-
-    expect(mergeEvents).toHaveLength(2)
-    expect(mergeEvents[0]?.metadata).toMatchObject({
-      decision: 'merged',
-      sourceUserId: source.user.id,
-      movedCredentialIds: [sourceCredential.id],
-      revokedSessionIds: [source.session.id],
-    })
-    expect(mergeEvents[1]?.metadata).toMatchObject({
-      decision: 'already-merged',
-      sourceUserId: source.user.id,
-      movedIdentityIds: [],
-      movedCredentialIds: [],
-      revokedSessionIds: [],
-    })
-    expect(JSON.stringify(mergeEvents)).not.toContain('source-secret')
-    expect(JSON.stringify(mergeEvents)).not.toContain('source@example.com')
-  })
-
-  it('rejects merge credential conflicts without moving data or leaking subjects into audit metadata', async () => {
-    const kit = createInMemoryAuthKit({
-      policy: createDefaultAuthPolicy({ allowMergeAccounts: true, requireReAuthFor: [] }),
-    })
-    const source = await kit.service.signIn({
-      assertion: assertion({
-        providerUserId: 'source-conflict',
-        email: 'source-conflict@example.com',
-      }),
-      now,
-    })
-    const target = await kit.service.signIn({
-      assertion: assertion({
-        providerUserId: 'target-conflict',
-        email: 'target-conflict@example.com',
-      }),
-      now,
-    })
-    const sourceCredential = await kit.service.setPassword({
-      userId: source.user.id,
-      email: 'source-conflict@example.com',
-      password: 'source-conflict-secret',
-      now,
-    })
-    await kit.service.setPassword({
-      userId: target.user.id,
-      email: 'target-conflict@example.com',
-      password: 'target-conflict-secret',
-      now,
-    })
-
-    const mergeError = await kit.service
-      .mergeAccounts({
-        sourceUserId: source.user.id,
-        targetUserId: target.user.id,
-        reAuthenticatedAt: now,
-        now,
-      })
-      .catch((caught: unknown) => caught)
-
-    expect(mergeError).toMatchObject({ code: UniAuthErrorCode.CredentialAlreadyExists })
-    expect(
-      kit.store
-        .listCredentials()
-        .filter((credential) => credential.id === sourceCredential.id)
-        .map((credential) => credential.userId),
-    ).toEqual([source.user.id])
-    expect(
-      kit.store.listUsers().find((user) => user.id === source.user.id)?.disabledAt,
-    ).toBeUndefined()
-    expect(
-      kit.store
-        .listSessions()
-        .filter((session) => session.userId === source.user.id)
-        .map((session) => session.status),
-    ).toEqual([SessionStatus.Active])
-
-    const denialEvent = [...kit.store.listAuditEvents()]
-      .reverse()
-      .find((event) => event.type === AuditEventType.PolicyDenied)
-
-    expect(denialEvent?.type).toBe(AuditEventType.PolicyDenied)
-    expect(denialEvent?.metadata).toMatchObject({
-      reason: 'merge-credential-conflict',
-      sourceUserId: source.user.id,
-      credentialTypes: ['password'],
-    })
-    expect(JSON.stringify(denialEvent?.metadata)).not.toContain('source-conflict-secret')
-    expect(JSON.stringify(denialEvent?.metadata)).not.toContain('source-conflict@example.com')
-    expect(JSON.stringify(denialEvent?.metadata)).not.toContain('target-conflict@example.com')
-  })
-
-  it('resolves assertions through a provider registry when requested', async () => {
-    const { providerRegistry, service } = createInMemoryAuthKit()
-    const provider = new StaticAuthProvider('telegram', {
-      providerUserId: 'pending',
-      displayName: 'Pending User',
-    })
-
-    provider.setAssertion({
-      providerUserId: 'tg-1',
-      displayName: 'Telegram User',
-    })
-    providerRegistry.register(provider)
-
-    const result = await service.signIn({
-      provider: 'telegram',
-      finishInput: { payload: { initData: 'signed' } },
-      now,
-    })
-
-    expect(result.identity.provider).toBe('telegram')
-    expect(result.user.displayName).toBe('Telegram User')
-  })
-})
+import './core/read-side-and-sessions.test.js'
+import './core/otp-and-verifications.test.js'
+import './core/policies-and-merge.test.js'
